Refactor part-one blockchain module
This is commit addresses a few concerns that were difficult to decompose.

It adds signatures to the Transactions, using the signer's public key as the source
address for withdrawing funds.

It seperates all mining concerns into a separate module, which will be extra credit.
This includes separate tests for these mineable classes.

The `isChainValid` method has been removed, and will be part of a seperate validation
module. You can't really trust a blockchain to tell you itself whether or not it is
valid.

Some general renaming and rewording.

Signed-off-by: Zac Delventhal <zac@bitwise.io>