Remove utils modules from part-two/client
There was a lot of defunct code in the source/utils/ directory, with a handful
of methods and constant that were used elsewhere. Splitting these off and
hard-coding them into the files that use them make the code a little less DRY,
but simplifies the directory structure, clarifying modules for students.

Signed-off-by: Zac Delventhal <zac@bitwise.io>