Add problemify script
This script is based on: https://github.com/bcmarinacci/problemify

It modifies JS source code in place, swapping in either problem or solution
code depending on the mode specified (--problem or --solution). Code is
designated using comments like on @bcmarinacci's problemify, with the
addition of /* PROBLEM FILE */ and /* SOLUTION FILE */ in the first line
of a file. These will result in the whole file being deleted in the
appropriate mode.

Source files will be stored on a staging branch, modified using this script
and then pushed to either `master` or `solution` branches.

Signed-off-by: Zac Delventhal <zac@bitwise.io>